feat: clean arch refactor
